requirements.txt
Flask==3.0.3
numpy==1.25.2
opencv_python==4.10.0.84
Pillow==10.0.0
scikit_learn==1.5.2
tensorflow==2.17.0
tensorflow_intel==2.17.0
